+++ app/lib/service/services.dart
@@ -156,8 +156,8 @@ Future<void> _withPubServices(FutureOr<void> Function() fn) async {
       ),
     );
     registerDartdocClient(DartdocClient());
-    registerDartSdkIndex(
-        InMemoryPackageIndex.sdk(urlPrefix: dartSdkMainUrl(toolEnvSdkVersion)));
+    registerDartSdkIndex(InMemoryPackageIndex.sdk(
+        urlPrefix: dartSdkMainUrl(toolStableSdkVersion)));
     registerJobBackend(JobBackend(dbService));
     registerNameTracker(NameTracker(dbService));
     registerPackageIndex(InMemoryPackageIndex());