+++ app/lib/shared/versions.dart
@@ -21,9 +21,9 @@ final RegExp runtimeVersionPattern = RegExp(r'^\d{4}\.\d{2}\.\d{2}$');
 /// Make sure that at least two versions are kept here as the next candidates
 /// when the version switch happens.
 const acceptedRuntimeVersions = <String>[
-  '2021.01.29', // The current [runtimeVersion].
+  '2021.02.05', // The current [runtimeVersion].
+  '2021.01.29',
   '2021.01.26',
-  '2021.01.18',
 ];
 
 /// Represents a combined version of the overall toolchain and processing,
@@ -53,7 +53,8 @@ bool shouldGCVersion(String version) =>
 
 // keep in-sync with SDK version in .mono_repo.yml and Dockerfile
 final String runtimeSdkVersion = '2.10.0';
-final String toolEnvSdkVersion = '2.12.0-133.2.beta';
+final String toolStableSdkVersion = '2.10.5';
+final String toolPreviewSdkVersion = '2.12.0-133.2.beta';
 
 // Value comes from package:pana.
 final String panaVersion = pana.packageVersion;